Created script to monitor changes to mse-set files
Script monitors create/change/rename/delete of '*.mse-set' files.
When a file is created/changed/renamed the script extracts the 'set' file into a similarly named directory.
When a file is deleted, the corresponding directory is deleted.
This should allow git diffs to compare changes in actual card fields, since mse-set files are binary and can't normally be diffed